novo-projeto/ativo-real/api/requirements_cosmosdb.txt
azure-functions==1.18.0
azure-cosmos==4.5.1
redis==5.0.1
